requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
dataclasses-json>=0.5.7
requests>=2.31.0
typing-extensions>=4.7.0
pytest>=7.0.0
marshmallow>=3.19.0 